requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h2bbff1b_1
astunparse=1.6.3=pypi_0
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=4.0.0=pyhd3eb1b0_0
ca-certificates=2021.7.5=haa95532_1
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py37haa95532_0
cffi=1.14.6=py37h2bbff1b_0
charset-normalizer=2.0.6=pypi_0
clang=5.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.2.0=h608a323_8
cycler=0.10.0=pypi_0
debugpy=1.4.1=py37hd77b12b_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
flatbuffers=1.12=pypi_0
gast=0.4.0=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.40.0=pypi_0
h5py=3.1.0=pypi_0
idna=3.2=pypi_0
importlib-metadata=4.8.1=py37haa95532_0
importlib_metadata=4.8.1=hd3eb1b0_0
ipykernel=6.2.0=py37haa95532_1
ipython=7.26.0=py37hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py37haa95532_1
jinja2=3.0.1=pyhd3eb1b0_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.7.1=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
keras=2.6.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py37h2bbff1b_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37hfa6e2cd_1001
msys2-conda-epoch=20160418=1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.3=py37haa95532_0
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
opencv-python=4.5.3.56=pypi_0
openssl=1.1.1l=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=21.0=pyhd3eb1b0_0
pandocfilters=1.4.3=py37haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.2=pypi_0
pip=21.0.1=py37haa95532_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyhca03da5_0
protobuf=3.18.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py37he774522_0
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pywin32=228=py37hbaba5e8_1
pywinpty=0.5.7=py37_0
pyzmq=22.2.1=py37hd77b12b_1
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.7.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=58.0.4=py37haa95532_0
six=1.15.0=pypi_0
sqlite=3.36.0=h2bbff1b_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow-estimator=2.6.0=pypi_0
tensorflow-gpu=2.6.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py37haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
tornado=6.1=py37h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=pypi_0
zipp=3.5.0=pyhd3eb1b0_0
